requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
attrs==25.3.0
azure-core==1.35.1
azure-identity==1.25.0
azure-storage-blob==12.26.0
backoff==2.2.1
boto3==1.36.0
botocore==1.36.26
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
croniter==6.0.0
cryptography==46.0.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.115.14
fastapi-sso==0.16.0
fastuuid==0.12.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.0
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
litellm==1.77.3
litellm-enterprise==0.1.20
litellm-proxy-extras==0.2.19
madoka==0.7.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mcp==1.12.4
mdurl==0.1.2
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
oauthlib==3.3.1
openai==1.108.1
orjson==3.11.3
packaging==25.0
polars==1.33.1
pondpond==1.4.1
propcache==0.3.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.18
pytz==2025.2
pywin32==311
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==13.7.1
rpds-py==0.27.1
rq==2.6.0
s3transfer==0.11.3
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.46.2
tiktoken==0.11.0
tokenizers==0.22.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.29.0
websockets==13.1
yarl==1.20.1
zipp==3.23.0
